chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/no-browsing-context-expected.txt

This ensures that @import doesn't work with headless documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL xhr.response.getElementById('hello').import should be null (of type object). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
